pyproject.toml
[project]
name = "bctpy-mrtrix"
version = "0.1.0"
description = "Brain Connectivity Toolbox analysis with MRTrix connectivity matrices"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "bctpy>=0.5.2",
    "scipy>=1.7.0",
    "statsmodels>=0.13.0",
    "openpyxl>=3.0.0",
]
